Sane+GObject: Implements get_scan_parameters()

+++ libinsane-gobject/include/libinsane-gobject/scan_parameters.h
@@ -38,7 +38,7 @@ LibinsaneImgFormat libinsane_scan_parameters_get_format(LibinsaneScanParameters 
 
 int libinsane_scan_parameters_get_width(LibinsaneScanParameters *self);
 int libinsane_scan_parameters_get_height(LibinsaneScanParameters *self);
-int libinsane_scan_parameters_get_image_size(LibinsaneScanParameters *self);
+gsize libinsane_scan_parameters_get_image_size(LibinsaneScanParameters *self);
 
 G_END_DECLS
